chore: wip refactor
The idea here was to reuse the experimental router within the actual
router. This turns out to be a lot of work without any security of
having something working and without breaking changes in the end. So I
think it's better to keep two versions of the createRouter function with
prefix `EXPERIMENTAL_`. In the end, one code base only uses one of the
function so it's fine to keep the code duplicated until v5. This branch
is here as a reminder of the failure.